pyproject.toml
[project]
name = "netsuite-shim"
version = "0.1.0"
description = "A thin Python client for NetSuite REST APIs, SuiteQL, and metadata discovery"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "httpx>=0.27,<1",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "PyJWT>=2.8,<3",
    "cryptography>=41,<44",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "respx>=0.21,<1",
    "ruff>=0.4,<1",
    "mypy>=1.10,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/netsuite_shim"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 99
